multiple dispatch and params prototype
