+++ _ old_jaxdf/__init__.py
@@ -1,0 +1,9 @@
+from jaxdf.core import *
+from jaxdf import discretization
+from jaxdf import geometry
+from jaxdf import ode
+from jaxdf import operators
+from jaxdf import primitives
+from jaxdf import spectral
+from jaxdf import utils
+from jaxdf import version
