+++ _ old_jaxdf/operators/__init__.py
@@ -1,0 +1,17 @@
+from arithmetic import *
+from jaxdf.discretization import *
+from jaxdf.core import bind_primitive
+
+
+@bind_primitive
+def add(x: Arbitrary, y: Arbitrary) -> Arbitrary:
+    # Process parameters
+    new_params = [x.params, y.params]
+
+    # Process get_field
+    def get_field(p, x):
+        return x.discretization.get_field(p, x) + y.discretization.get_field(p, y)
+
+    return None, Arbitrary(
+        x.domain, get_field, 
+    )
