+++ jaxdf/core.py
@@ -1,375 +1,70 @@
-from jaxdf.geometry import Domain
-from jaxdf import operators as jops
-from typing import List, Any, Set
-from collections.abc import Iterable
-from hashids import Hashids
+
+import inspect
+from functools import partial, wraps
+from typing import get_type_hints, Any, NewType, TypeVar
+
+from jax.tree_util import register_pytree_node, register_pytree_node_class
+
 from plum import dispatch
-from functools import wraps
 
 
-class Discretization(object):
-    domain: Domain
-
-    def init_params(self, seed):
-        return 0
-
-    def __repr__(self):
-        return self.__class__.__name__
-
-
-class Globals(object):
-    def __init__(self):
-        self.dict = {"shared": {}, "independent": {}}
-
-    def get(self, group=None, key=None):
-        if group is None and key is None:
-            return self.dict
-        return self.dict[group][key]
-
-    def set(self, key: str, value: Any, group: str):
-        if group == "shared":
-            for k, v in value.items():
-                if k not in self.dict["shared"].keys():
-                    self.dict["shared"][k] = v
-        elif group == "independent":
-            if key not in self.dict["independent"].keys():
-                self.dict["independent"][key] = value
-            else:
-                raise ValueError(f"{key} already exists in independent Globals!")
-
-    def __repr__(self):
-        str1 = f"Shared: {self.dict['shared']}\n"
-        str2 = f"Independent: {self.dict['independent']}"
-        return str1 + str2
-
-
-class Tracer(object):
-    def __init__(self):
-        self.input_fields = {}
-        self.globals = Globals()
-        self.operations = {}
-        self._counter = 0
-        self._hash = Hashids()
-
-    def counter(self):
-        self._counter += 1
-        return self._hash.encode(self._counter)
-
-    def add_input_field(self, name, field):
-        if name not in self.input_fields.keys():
-            self.input_fields[name] = field
-
-    def add_operation(self, op):
-        self.operations[op.yelds.name] = op
-
-    def construct_function(self, sorted_graph: dict, names: List[str]):
-        concrete_ops = {}
-
-        def f(global_params, field_params):
-            # Add used inputs
-            for field_name in field_params.keys():
-                if field_name in sorted_graph:
-                    concrete_ops[field_name] = field_params[field_name]
-
-            # Compose operations present in the sorted grap
-            for op_name in self.operations.keys():
-                if op_name in sorted_graph:
-                    op = self.operations[op_name]
-                    args = [concrete_ops[n] for n in op.inputs]
-                    if op.param_kind == "none":
-                        op_params = {}
-                    elif op.param_kind == "shared":
-                        op_params = global_params["shared"]
-                    else:
-                        op_params = global_params[op.param_kind][op.params]
-                    concrete_ops[op_name] = op.fun(op_params, *args)
-
-            # Output requested parameters
-            outputs = []
-            for n in names:
-                outputs.append(concrete_ops[n])
-            return outputs
-
-        return f
-
-    @staticmethod
-    def _filter_unneded_ops(operations: dict, out_fields: List[str], debug=False) -> Set:
-        op_keys = [k for k in operations.keys()]
-        op_args = [k.inputs for k in operations.values()]
-        stack = out_fields
-        out_set : List = []
-        while stack:
-            if debug:
-                print(out_set)
-            x = stack.pop()
-            out_set.append(x)
-            # if is not an input or constant
-            if x in op_keys:
-                for x in op_args[op_keys.index(x)]:
-                    if x not in out_set:
-                        stack.append(x)
-        return set(out_set)
-
-    def __repr__(self):
-        ops = ["- " + str(x) + "\n" for x in self.operations.values()]
-        ops = "".join(ops)
-        input_keys = tuple([*self.input_fields.keys()])
-        string = f"Input fields: {input_keys}\n\n"
-        string += f"Globals: {self.globals}\n\n"
-        string += f"Operations:\n{ops}"
-        return string
-
-
-def is_numeric(x):
-    return (type(x) == int) + (type(x) == float) + (type(x) == complex)
-
-
-class Field(object):
-    def __init__(self, discretization, name, params):
-        self.params = params
-        self.discretization = discretization
-        self.name = name
-
-    def __call__(self, x):
-        return self.discretization.get_field()(self.params, x)
-
-    def get_field(self):
-        return self.discretization.get_field()
-
-    def init_params(self, seed):
-        return self.discretization.init_params(seed)
-
-    def __add__(self, other):
-        if is_numeric(other):
-            return jops.add_scalar(self, other)
-        else:
-            return jops.add(self, other)
-
-    def __truediv__(self, other):
-        if is_numeric(other):
-            return jops.div_scalar(self, other)
-        else:
-            return jops.div(self, other)
-
-    def __mul__(self, other):
-        if is_numeric(other):
-            return jops.mul_scalar(self, other)
-        else:
-            return jops.mul(self, other)
-
-    def __pow__(self, other):
-        if is_numeric(other):
-            return jops.power_scalar(self, other)
-        else:
-            return jops.power(self, other)
-
-    def __radd__(self, other):
-        return self + other
-
-    def __rmul__(self, other):
-        return self * other
-
-    def __rtruediv__(self, other):
-        self = jops.reciprocal(self)
-        return self * other
-
-    def __rsub__(self, other):
-        self = jops.invert(self)
-        return self + other
-
-    def __sub__(self, other):
-        if is_numeric(other):
-            return jops.add_scalar(self, -other)
-        else:
-            other = jops.invert(other)
-            return self + other
-
-    def __repr__(self):
-        return f"Field :: {self.discretization}, {self.name}"
-
-
-class TracedField(Field):
-    def __init__(
-        self, discretization: Discretization, params: dict, tracer: Tracer, name: str
-    ):
-        self.params = params
-        self.discretization = discretization
-        self.tracer = tracer
-        self.name = name
-        self.preprocess_params = lambda x, y: y  # Identity function for parameters
-
-    @staticmethod
-    def from_field(field, tracer, name):
-        return TracedField(field.discretization, field.params, tracer, name)
-
-    def __repr__(self):
-        return f"-- TracedField :: {self.discretization}, {self.name} \n\n{self.tracer}"
-
-    def get_field(self):
-        f = self.discretization.get_field()
-
-        def wrapped_f(global_params, input_params, x):
-            new_params = self.preprocess_params(global_params, input_params)
-            return f(new_params, x)
-
-        wrapped_f.__name__ = f.__name__
-        return wrapped_f
-
-    def get_field_on_grid(self):
-        f = self.discretization.get_field_on_grid()
-
-        def wrapped_f(global_params, input_params):
-            new_params = self.preprocess_params(global_params, input_params)
-            return f(new_params)
-
-        wrapped_f.__name__ = f.__name__
-        return wrapped_f
-
-
-def evaluate(op):
-    @operator(debug=False)
-    def _eval(*args, **kwargs):
-        return op(*args, **kwargs)
-
-    return _eval
-
-
-class DiscretizedOperator(object):
-    def __init__(
-        self, global_params, preprocess_func, fields, global_preprocess, tracer
-    ):
-        self.globals = global_params
-        self.preprocess_func = preprocess_func
-        self.fields = fields
-        self.global_preprocess = global_preprocess
-        self.tracer = tracer
-
-    def __repr__(self):
-        discr = [x.discretization for x in self.fields]
-        names = [x.name for x in self.fields]
-        return f"DiscretizedOperator :: {discr}, {names} \n\n {self.tracer}"
-
-    def get_global_params(self):
-        return self.tracer.globals.get()
-
-    def get_field(self, idx=-1):
-        """idx=-1 means all fields"""
-        if idx != -1:
-            f = self.fields[idx].discretization.get_field()
-            preprocess = self.preprocess_func[idx]
-
-            def wrapped_f(global_params, input_params, x):
-                new_params = preprocess(global_params, input_params)
-                return f(new_params, x)
-
-            return wrapped_f
-        else:
-            preprocess = self.global_preprocess
-            f_all = [f.discretization.get_field() for f in self.fields]
-
-            def wrapped_f(global_params, input_params, x):
-                all_new_params = preprocess(global_params, input_params)
-                return [f(all_new_params[i], x) for i, f in enumerate(f_all)]
-
-            return wrapped_f
-
-    def get_field_on_grid(self, idx=-1):
-        """idx=-1 means all fields"""
-        if idx != -1:
-            f = self.fields[idx].discretization.get_field_on_grid()
-            preprocess = self.preprocess_func[idx]
-
-            def wrapped_f(global_params, input_params):
-                new_params = preprocess(global_params, input_params)
-                return f(new_params)
-
-            return wrapped_f
-        else:
-            preprocess = self.global_preprocess
-            f_all = [f.discretization.get_field_on_grid() for f in self.fields]
-
-            def wrapped_f(global_params, input_params):
-                all_new_params = preprocess(global_params, input_params)
-                return [f(all_new_params[i]) for i, f in enumerate(f_all)]
-
-            return wrapped_f
-
-def bind_primitive(f):
-    return wraps(f)(dispatch(f))
-
-
-
-def operator(has_aux=False, debug=False):
-    """Returns a decorator that builds the computational graph of an operator"""
-    if has_aux is True:
-        raise NotImplementedError(
-            "Operator with auxiliary arguments are not currently supported."
-        )
-
-    def decorator(fun):
-        def wrapper(*args, **kwargs):
-
-            tracer = Tracer()
-
-            # Use only named variables
-            if len(args) > 0:
-                raise ValueError(
-                    "Only keyword arguments can be used in a traced function"
-                )
-
-            # Register input fields
-            for input_name in kwargs.keys():
-                field = TracedField.from_field(
-                    kwargs[input_name], tracer, name=input_name
-                )
-                tracer.add_input_field(name=input_name, field=field)
-
-            # Register operations
-            out_fields = fun(*[x for x in tracer.input_fields.values()])
-            if not isinstance(out_fields, Iterable):
-                out_fields = (out_fields,)
-
-            # Register discretization parameters
-            """
-            for f in out_fields:
-                tracer.add_discretization(f.discretization)
-            """
-
-            # Generates jax function for each output variable
-            out_names = [x.name for x in out_fields]
-            p_func = []
-            fields = []
-            for out in out_names:
-                sorted_graph = tracer._filter_unneded_ops(
-                    tracer.operations, [out], debug=debug
-                )
-                if debug:
-                    print(tracer.operations)
-                    print(f"Sorted graph for {out}:\n{sorted_graph}\n")
-                pf = tracer.construct_function(sorted_graph, [out])
-
-                def preprocess_func(x, y):
-                    return pf(x, y)[0]  # Only need one output
-
-                field = tracer.operations[out].yelds
-
-                p_func.append(preprocess_func)
-                fields.append(field)
-
-            # Making function preprocessing parameters for all fields
-            sorted_graph = tracer._filter_unneded_ops(tracer.operations, out_names)
-            global_preprocess = tracer.construct_function(sorted_graph, out_names)
-
-            output = DiscretizedOperator(
-                global_params=tracer.globals,
-                preprocess_func=p_func,
-                fields=fields,
-                global_preprocess=global_preprocess,
-                tracer=tracer,
-            )
-
-            return output
-
-        return wrapper
-
-    return decorator
+Params = None
+@register_pytree_node_class
+class Operator:
+    def __init__(self, setup_fun, evaluate):
+        self.setup_fun = setup_fun
+        self.evaluate = evaluate
+
+    def __call__(self, *args, **kwargs):
+        # Check if tracer exists in any of the arguments and
+        # it is not None.
+        for arg in args:
+            if hasattr(arg, "_tracer") and arg._tracer is not None:
+                return self._traced_call(*args, **kwargs)
+        for arg in kwargs.values():
+            if hasattr(arg, "_tracer") and arg._tracer is not None:
+                return self._traced_call(*args, **kwargs)
+        return self._call(*args, **kwargs)
+
+    def _traced_call(self, *args, **kwargs):
+        pass
+
+    def _call(self, *args, **kwargs):
+        # Construct the operator parameters
+        op_params = self.setup_fun(*args, **kwargs)
+        # Runs the parametrized function
+        field = self.evaluate(*args, **kwargs, params = op_params)
+        return field
+
+    def tree_flatten(self):
+        children = {}
+        aux_data = (self.setup_fun, self.evaluate)
+        return (children, aux_data)
+
+    @classmethod
+    def tree_unflatten(cls, aux_data, children):
+        return cls(setup_fun=aux_data[0], evaluate=aux_data[1])
+
+def _operator(evaluate, setup_fun):
+  op = Operator(setup_fun, evaluate)
+  
+  @wraps(evaluate)
+  def _operator(*args, **kwargs):
+    return op(*args, **kwargs)
+
+  return dispatch(_operator)
+
+def operator(evaluate):
+  setup_fun = lambda *args, **kwargs: None
+    
+  @wraps(evaluate)
+  def evaluate_wrapper(*args, params, **kwargs):
+    return evaluate(*args, **kwargs)
+    
+  return _operator(evaluate_wrapper, setup_fun)
+
+def parametrized_operator(setup_fun):
+  def decorate(evaluate):
+    op = _operator(evaluate, setup_fun)
+    return op
+  return decorate
